Fix Apple watch items customization not sync in the same session
